Move from bundled types and extensions to EnumerableToolkit
