+++ MonkeyLoader/Logging/LoggingHandler.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using EnumerableToolkit;
+using System;
 using System.Collections.Generic;
 using System.Linq;
