+++ MonkeyLoader/Meta/ModLoadingLocation.cs
@@ -1,4 +1,5 @@
-﻿using Newtonsoft.Json;
+﻿using EnumerableToolkit;
+using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
