+++ MonkeyLoader/Patching/MonkeyBase.PatchJob.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Diagnostics.CodeAnalysis;
+using EnumerableToolkit;
 
 namespace MonkeyLoader.Patching
 {
