+++ MonkeyLoader/Patching/PrePatchTarget.cs
@@ -1,4 +1,5 @@
-﻿using Mono.Cecil;
+﻿using EnumerableToolkit;
+using Mono.Cecil;
 using System;
 using System.Collections.Generic;
 using System.Linq;
